+++ source/Contracts/Document.cs
@@ -59,6 +59,6 @@ namespace DreadBot
 		/// Optional. File size
 		/// </summary>
 		[DataMember(Name = "file_size", EmitDefaultValue = false)]
-		public int file_size { get; set; }
+		public long file_size { get; set; }
 	}
 }
