+++ source/Contracts/Video.cs
@@ -69,6 +69,6 @@ namespace DreadBot
 		/// Optional. File size
 		/// </summary>
 		[DataMember(Name = "file_size", EmitDefaultValue = false)]
-		public int file_size { get; set; }
+		public long file_size { get; set; }
 	}
 }
